fix: make error handle unknown types
The Wyre documentation implies there is a fixed set of error types, but
the documented list is non-exhaustive. We've encountered at least two
error types that are not on the list, and there are likely more. This
commit avoids the problem by using a String instead of an enum so it
can vacuously accept any error type. For matching and documentation's
sake we still include the known error types as consts.

This also addresses a semi-related issue, where the ApiError fields
were private. They need to be public so the error can be inspected and
handled properly.

This also adds the Serialize trait to a few types to aid in debugging.